+++ src/containers/ChartViewContainer.tsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { Dispatch } from 'redux';
 
-import * as actions from '../actions';
+import actions from '../actions';
 import ChartView from '../components/ChartView';
 import { Chart, StoreState } from '../store/types';
 import { push } from 'react-router-redux';
@@ -26,9 +26,9 @@ function mapStateToProps({ charts }: StoreState, { match: { params } }: RoutePro
 
 function mapDispatchToProps(dispatch: Dispatch<StoreState>) {
   return {
-    getChart: (id: string) => dispatch(actions.getChart(id)),
+    getChart: (id: string) => dispatch(actions.charts.getChart(id)),
     deployChart: (chart: Chart, releaseName: string, namespace: string) =>
-      dispatch(actions.deployChart(chart, releaseName, namespace)),
+      dispatch(actions.charts.deployChart(chart, releaseName, namespace)),
     push: (location: string) => dispatch(push(location)),
   };
 }
