+++ src/reducers/index.ts
@@ -1,6 +1,7 @@
 import { getType } from 'typesafe-actions';
 
-import * as actions from '../actions/index';
+import actions from '../actions';
+import { ChartsAction } from '../actions/charts';
 import { ChartState } from '../store/types';
 
 const initialState: ChartState = {
@@ -10,14 +11,14 @@ const initialState: ChartState = {
   items: [],
 };
 
-function charts(state: ChartState = initialState, action: actions.RootAction): ChartState {
+function charts(state: ChartState = initialState, action: ChartsAction): ChartState {
   switch (action.type) {
-    case getType(actions.requestCharts):
-      return {...state, isFetching: true};
-    case getType(actions.receiveCharts):
-      return {...state, isFetching: false, items: action.charts};
-    case getType(actions.selectChart):
-      return {...state, isFetching: false, selectedChart: action.chart};
+    case getType(actions.charts.requestCharts):
+      return { ...state, isFetching: true };
+    case getType(actions.charts.receiveCharts):
+      return { ...state, isFetching: false, items: action.charts };
+    case getType(actions.charts.selectChart):
+      return { ...state, isFetching: false, selectedChart: action.chart };
     default:
   }
   return state;
